+++ src/App.js
@@ -5,12 +5,14 @@ import Footer from './Components/Footer';
 import Header from './Components/Header';
 import ItemMenus from './Components/ItemMenus';
 import Statement from './Components/Statement';
+import StatementHeading from './Components/StatementHeading';
 
 function App() {
   return (
     <div>
       {/* <Anc></Anc> */}
       <Header></Header>
+      <StatementHeading></StatementHeading>
       <Nav></Nav>
       <Statement></Statement>
       <ItemMenus></ItemMenus>
